requirements.txt
jax>=0.4.0
jaxlib>=0.4.0
distrax>=0.1.2
optax>=0.1.4
matplotlib>=3.6.2
pyscf>=2.1.1
dm-haiku>=0.0.9
chex==0.1.5
pandas>=1.5.2
mpmath>=1.2.1
tensorflow>=2.11.0
pubchempy==1.0.4
tqdm
matplotlib
pandas
einops
jaxtyping
